+++ cypress/support/graphql_testcase.js
@@ -1,10 +1,19 @@
 export function version() {
+  const query = 'query' + '{version}'
+  cy.addGraphqlLogs(query)
+
   return {
     query: 'query' + '{version}',
     queryVariables: {},
   }
 }
 export function viewList({ shopperId, name }) {
+  const query =
+    'query' +
+    '($shopperId: String!,$name: String!)' +
+    '{  viewList(shopperId:$shopperId,name:$name) {public,name,range {total,from,to},data {id,productId,sku,title}}}'
+
+  cy.addGraphqlLogs(query, { shopperId, name })
   return {
     query:
       'query' +
@@ -18,6 +27,11 @@ export function viewList({ shopperId, name }) {
 }
 
 export function viewLists() {
+  const query =
+    'query' +
+    '($shopperId: String!)' +
+    '{  viewLists(shopperId:$shopperId) {public,name,range {total,from,to},data {id,productId,sku,title}}}'
+  cy.addGraphqlLogs(query, 'saravananvenkatesan@bitcot.com')
   return {
     query:
       'query' +
@@ -30,6 +44,15 @@ export function viewLists() {
 }
 
 export function checkList() {
+  const query =
+    'query' +
+    '($shopperId: String!,$productId:String!)' +
+    '{  checkList(shopperId:$shopperId,productId:$productId) {inList,listNames,listIds,message}}'
+  const queryVariables = {
+    shopperId: 'saravananvenkatesan@bitcot.com',
+    productId: '880300',
+  }
+  cy.addGraphqlLogs(query, queryVariables)
   return {
     query:
       'query' +
@@ -43,6 +66,14 @@ export function checkList() {
 }
 
 export function listNames() {
+  const query =
+    'query' + '($shopperId: String!)' + '{  listNames(shopperId:$shopperId)}'
+  const queryVariables = {
+    shopperId: 'saravananvenkatesan@bitcot.com',
+  }
+
+  cy.addGraphqlLogs(query, queryVariables)
+
   return {
     query:
       'query' + '($shopperId: String!)' + '{  listNames(shopperId:$shopperId)}',
@@ -58,6 +89,7 @@ export function addToList(productId) {
     '($shopperId: String!, $listItem: ListItemInputType!, $name: String!)' +
     '{addToList(shopperId: $shopperId, listItem: $listItem, name: $name)}'
 
+  cy.addGraphqlLogs(query, productId)
   return {
     query,
     queryVariables: productId,
@@ -70,6 +102,8 @@ export function removeFromList(productId, { shopperId, name }) {
     '($shopperId: String!,$id: ID!,$name: String)' +
     '{removeFromList(shopperId: $shopperId,id: $id,name: $name)}'
 
+  cy.addGraphqlLogs(query, productId)
+
   return {
     query,
     queryVariables: {
@@ -78,6 +112,29 @@ export function removeFromList(productId, { shopperId, name }) {
       name,
     },
   }
+}
+
+export function exportList(email) {
+  const query =
+    'query' +
+    '{ exportList{email,listItemsWrapper{listItems{id,productId,sku,title},isPublic}}}'
+
+  cy.addGraphqlLogs(query, email)
+  return {
+    query:
+      'query' +
+      '{ exportList{email,listItemsWrapper{listItems{id,productId,sku,title},isPublic}}}',
+    queryVariables: {
+      email,
+    },
+  }
+}
+
+export function validateexportListResponse(response) {
+  expect(response.body.data).to.not.equal(null)
+  expect(response.body.data.exportList[0].email).to.be.equal(
+    'saravananvenkatesan@bitcot.com'
+  )
 }
 
 export function validateGetVersionResponse(response) {